% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoSample}
\alias{demoSample}
\title{Sample demographics dataset}
\format{
A data table with 5 rows and the 5 following columns:

\describe{
\item{\code{person_id}}{Character vector of the identifier for each
person in the cohort}
\item{\code{sex}}{Character vector indicating biological sex
('M': male, 'F': female)}
\item{\code{uniq_age}}{Numeric vector containing the number of unique
years with ICD codes}
\item{\code{first_age}}{Numeric vector containing the age (in days) when
the person first received an ICD code}
\item{\code{last_age}}{Numeric vector containing the age (in days) when
the person last received an ICD code}
}
}
\usage{
demoSample
}
\description{
This table contains a sample dataset of patient demographics information.
}
\seealso{
\code{\link[=calcWeights]{calcWeights()}}, \code{\link[=calcPheRS]{calcPheRS()}}
}
\keyword{datasets}
