% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diseaseDxPhecode}
\alias{diseaseDxPhecode}
\title{Diagnostic phecodes}
\format{
A data table with 1 row and the 2 following columns:

\describe{
\item{\code{disease_id}}{Numeric vector of disease identifiers}
\item{\code{dx_phecode}}{Character vector of the phecode that represents
diagnosis with \code{disease_id}}
}
}
\source{
<>
}
\usage{
diseaseDxPhecode
}
\description{
This table provides phecodes that represent diagnosis with a Mendelian disease.
}
\seealso{
\code{\link[=runPositiveControl]{runPositiveControl()}}
}
\keyword{datasets}
