% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{diseaseHpoMap}
\alias{diseaseHpoMap}
\title{Mapping of Mendelian diseases and their clinical features}
\format{
A data table with 75,084 rows and the 6 following columns:
\itemize{
\item \code{disease_id}: Numeric vector of OMIM disease identifiers corresponding to
the disease being annotated
\item \code{disease_name}: Character vector of disease name
\item \code{hpo_term_id}: Character vector of the HPO identifier corresponding to the
phenotype mapped to \code{disease_id}
\item \code{hpo_term_name}: Character vector containing the description of \code{hpo_term_id}
}
}
\source{
\url{https://hpo.jax.org/app/download/annotation}
}
\usage{
diseaseHpoMap
}
\description{
This table provides mapping between Mendelian diseases and the clinical
features they are linked to using annotations from
Online Mendelian Inheritance in Man (OMIM).
}
\seealso{
\code{\link[=mapDiseaseToPhecode]{mapDiseaseToPhecode()}}
}
\keyword{datasets}
