% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{diseaseInfo}
\alias{diseaseInfo}
\title{Mendelian disease annotation}
\format{
A data table with 63 rows and the 3 following columns:
\itemize{
\item \code{disease_id}: Numeric vector of OMIM disease identifiers
\item \code{disease_name}: Character vector of disease names
\item \code{gene}: Character vector containing the gene known to cause the disease
}
}
\source{
<>
}
\usage{
diseaseInfo
}
\description{
This table provides annotations for Mendelian diseases
}
\seealso{
\code{\link[=getGeneticAssociations]{getGeneticAssociations()}}
}
\keyword{datasets}
