% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneVarMap}
\alias{geneVarMap}
\title{Mapping of genes and variants}
\format{
A data table with 65,808 rows and the 3 following columns:

\describe{
\item{\code{gene}}{Character vector of gene names}
\item{\code{vid}}{Character vector of variant IDs based on position on
the chromosome}
\item{\code{rsid}}{Character vector of dbSNP Reference SNP numbers (rs)}
}
}
\source{
\url{https://annovar.openbioinformatics.org/en/latest/}
}
\usage{
geneVarMap
}
\description{
This table provides mapping between genes and genetic variants using
ANNOVAR annotations
}
\seealso{
\code{\link[=genotypeAssociation]{genotypeAssociation()}}
}
\keyword{datasets}
