% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_association.R
\name{genotypeAssociation}
\alias{genotypeAssociation}
\title{Perform genotype association tests}
\usage{
genotypeAssociation(
  scores,
  genotypes,
  covariates,
  formStr = as.formula("score ~ ."),
  coding = "additive"
)
}
\arguments{
\item{scores}{A data.table of phenotype risk scores. The columns are
\code{person_id}, \code{disease_id}, \code{score}.}

\item{genotypes}{A data.table of genotypes. The columns are \code{person_id}, and
variant IDs to be tested.}

\item{covariates}{A data.table of covariates to be used in the association
analysis.}

\item{formStr}{a formula object representing the linear model for
genetic associations. must have 'score' as the dependent variable
and 'variant' as the independent variable. All other independent variables
must be existing columns in \code{covariates}.}

\item{coding}{coding of the genotypes. Either 'additive' or 'genotypic'}
}
\value{
A data.table containing to results of the genotype association tests
with the following columns. If the model fails to converge,
NAs will be reported.
\item{\code{disease_id}}{disease identifier}
\item{\code{vid}}{variant identifier}
\item{\code{beta}}{The beta coefficient for the variant}
\item{\code{se}}{The standard error for the beta coefficient}
\item{\code{p}}{The p-value for the variant}
\item{\code{lower}}{The lower bound of the \code{confint} confidence interval}
\item{\code{upper}}{The upper bound of the \code{confint} confidence interval}
}
\description{
Perform association tests between the phenotype risk scores and genotypes
}
