% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_control.R
\name{getCases}
\alias{getCases}
\title{Identify Mendelian disease cases}
\usage{
getCases(phecodes, diseaseDxPhecode = phers::diseaseDxPhecode)
}
\arguments{
\item{phecodes}{A data.table containing phenotypes stored as phecodes
for each person. The columns are \code{person_id} and \code{phecode}.}

\item{diseaseDxPhecode}{A data.table containing mapping between diseases and
the phecodes that represent being diagnosed with them. The columns are
\code{disease_id} and \code{dx_phecode}.}
}
\value{
A data.table with cases of each disease in \code{diseaseDxPhecode}.
The columns are \code{person_id}, \code{disease_id}.
}
\description{
Identify cases of Mendelian diseases.
}
