% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{hpoPhecodeMap}
\alias{hpoPhecodeMap}
\title{Mapping of HPO terms and phecodes}
\format{
A data table with 4189 rows and the 4 following columns:
\itemize{
\item \code{hpo_term_id}: Character vector representing the HPO term ID to be mapped
\item \code{hpo_term_name}: Character vector containing the description of \code{hpo_term_id}
\item \code{phecode}: Character vector representing the phecode mapping
\item \code{phecode_name}: Character vector containing the description of each phecode
}
}
\usage{
hpoPhecodeMap
}
\description{
This table provides mapping between Human Phenotype Ontology (HPO)
annotations and phecodes (groupings of ICD-9-CM and ICD-10-CM)
}
\seealso{
\code{\link[=mapDiseaseToPhecode]{mapDiseaseToPhecode()}}
}
\keyword{datasets}
