% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{icdSample}
\alias{icdSample}
\title{Sample diagnostic code dataset}
\format{
A data table with 141 rows and the 4 following columns:
\itemize{
\item \code{person_id}: Character vector of the identifier for each
person
\item \code{icd}: Character vector containing the ICD codes recorded
for each person
\item \code{flag}: Numeric vector representing the vocabulary of the
ICD code (\strong{9}: ICD-9-CM,  \strong{10}: ICD-10-CM)
\item \code{entry_date}: vector of type \code{Date} indicating the date ICD code was
recorded.
}
}
\usage{
icdSample
}
\description{
This table contains a sample dataset of patient diagnostic codes
(ICD-9-CM and ICD-10-CM).
}
\seealso{
\code{\link[=getPhecodeOccurrences]{getPhecodeOccurrences()}}, \code{\link[=getWeights]{getWeights()}}, \code{\link[=getScores]{getScores()}}
}
\keyword{datasets}
