% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{phers}
\alias{phers}
\title{Run through the workflow for calculating phenotype risk scores}
\usage{
phers(
  demos,
  icdOccurrences,
  diseasePhecodeMap,
  icdPhecodeMap = phers::icdPhecodeMap,
  dxIcd = phers::diseaseDxIcdMap,
  preCalcWeights = NULL,
  residScoreFormula = NULL
)
}
\arguments{
\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}.}

\item{icdOccurrences}{A data.table of occurrences of ICD codes for each
person in the cohort. Must have columns \code{person_id}, \code{icd}, and \code{flag}.}

\item{diseasePhecodeMap}{A data.table containing the mapping between
diseases and phecodes. Must have columns \code{disease_id} and \code{phecode}.}

\item{icdPhecodeMap}{A data.table containing the mapping between ICD codes
and phecodes. Must have columns \code{icd}, \code{phecode}, and \code{flag}. By default uses
the mapping included in this package.}

\item{dxIcd}{A data.table of diagnostic ICD codes to remove for every person.
Must have columns \code{icd} and \code{flag}. By default uses a table provided in
this package with mapping between diseases and ICD codes used to indicate
their diagnosis. If \code{NULL} no ICD codes will be removed.}

\item{preCalcWeights}{A data.table of pre-calculated weights.
May use weights provided in this package: \code{phers::preCalcWeights},
which are calculated using data from Vanderbilt University Medical Center.
Recommended when data provided by the user has low sample size.}

\item{residScoreFormula}{A formula object representing the linear model of
covariates to be used in the calculation of residual phenotype risk score.
If \code{NULL} residual scores won't be calculated.}
}
\value{
A data.table containing the phenotype risk score for each person for
each disease.
}
\description{
This function calculates phenotype risk scores given patient ICD codes
and demographic data without requiring intermediate steps for
calculating phecode occurrences and their weights.
}
\examples{
library('data.table')

# vector of OMIM disease IDs to calculate PheRS for
diseaseId = 154700

# mape diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()
diseasePhecodeMap = diseasePhecodeMap[disease_id \%in\% diseaseId]

# calculate PheRS
scores = phers(demoSample, icdSample, diseasePhecodeMap)

  
}
