% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{runPhers}
\alias{runPhers}
\title{Run through the workflow for calculating phenotype risk scores}
\usage{
runPhers(
  demos,
  icdOccurrences,
  diseaseIds,
  diseasePhecodeMap,
  icdPhecodeMap = phers::icdPhecodeMap,
  dxIcd = phers::diseaseDxIcdMap
)
}
\arguments{
\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}.}

\item{icdOccurrences}{A data.table of occurrences of ICD codes for each
person in the cohort. Must have columns \code{person_id}, \code{icd}, and \code{flag}.}

\item{diseaseIds}{A numeric vector of OMIM disease IDs to calculate
phenotype risk scores for.}

\item{diseasePhecodeMap}{A data.table containing the mapping between
diseases and phecodes. Must have columns \code{disease_id} and \code{phecode}.}

\item{icdPhecodeMap}{A data.table containing the mapping between ICD codes
and phecodes. Must have columns \code{icd}, \code{phecode}, and \code{flag}. By default uses
the mapping included in this package.}

\item{dxIcd}{A data.table of diagnostic ICD codes to remove for every person.
Must have columns \code{icd} and \code{flag}. By default uses a table provided in
this package with mapping between diseases and ICD codes used to indicate
their diagnosis. If \code{NULL} no ICD codes will be removed.}
}
\value{
A data.table containing the phenotype risk score for each person for
each disease.
}
\description{
This function calculates phenotype risk scores given patient ICD codes
and demographic data without requiring intermediate steps for
calculating phecode occurrences and their weights.
}
\examples{
library('data.table')

# vector of OMIM disease IDs to calculate PheRS for
diseaseId = 154700

# mape diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()

# calculate PheRS
scores = runPhers(demoSample, icdSample, diseaseId, diseasePhecodeMap)

  
}
