% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_control.R
\name{runPositiveControl}
\alias{runPositiveControl}
\title{Positive control analysis}
\usage{
runPositiveControl(scores, cases, diseaseID)
}
\arguments{
\item{scores}{A data.table of phenotype risk scores. The columns are
\code{person_id}, \code{disease_id}, \code{phers}.}

\item{cases}{A data.table with cases of each disease in \code{diseaseDxPhecode}.
The columns are \code{person_id}, \code{disease_id}.}

\item{diseaseID}{Numeric value of the disease identifier to generate
the plot for.}
}
\value{
A box plot showing the difference in distribution of cases and controls.
}
\description{
Perform positive control analysis
}
