% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{calcWeights}
\alias{calcWeights}
\title{Calculate phenotype weights}
\usage{
calcWeights(demos, phecodes)
}
\arguments{
\item{demos}{A data.table containing demographic information for each person
in the population. The columns are \code{person_ID}, \code{sex}, \code{uniq_age},
\code{first_age}, and \code{last_age}. The age columns specify the number of unique years with
ICD codes, the age of the individual at first ICD code,
and the age of the individual at last ICD code.}

\item{phecodes}{A data.table containing phenotypes stored as phecodes
for each person. The columns are \code{person_ID} and \code{phecode}.}
}
\value{
A data.table where each row is a phecode and the population
prevalence and weight corresponding to it.
The columns are \code{phecode}, \code{prev}, \code{w}.
}
\description{
Calculate the population prevalence and weight for each phenotype.
These weights will be used in the calculation of the phenotype risk score.
}
