% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{demoSample}
\alias{demoSample}
\title{Sample table of demographic information}
\format{
A data table with the following columns:
\itemize{
\item \code{person_id}: Character vector of the identifier for each person in the
cohort
\item \code{sex}: Character vector indicating biological sex
}
}
\usage{
demoSample
}
\description{
The data are artificial and do not correspond to real patients.
}
\seealso{
\code{\link[=getWeights]{getWeights()}}, \code{\link[=getScores]{getScores()}}
}
\keyword{datasets}
