% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{demoSample}
\alias{demoSample}
\title{Sample demographics dataset}
\format{
A data table with 5 rows and the 2 following columns:
\itemize{
\item \code{person_id}: Character vector of the identifier for each
person in the cohort
\item \code{sex}: Character vector indicating biological sex
}
}
\usage{
demoSample
}
\description{
This table contains a sample dataset of patient demographics information.
}
\seealso{
\code{\link[=getWeights]{getWeights()}}, \code{\link[=getScores]{getScores()}}
}
\keyword{datasets}
