% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{diseaseDxIcdMap}
\alias{diseaseDxIcdMap}
\title{Mapping of diseases and diagnostic ICD codes}
\format{
A data table with 63 rows and the 3 following columns:
\itemize{
\item \code{disease_id}: Numeric vector of OMIM disease identifiers
\item \code{disease_name}: Character vector of disease names
\item \code{icd}: Character vector representing the diagnostic ICD code
\item \code{icd_string}: Character vector containing the description of each icd code
\item \code{flag}: Integer vector representing the vocabulary of the ICD code
(\strong{9}: ICD-9-CM,  \strong{10}: ICD-10-CM)
}
}
\source{
\url{https://phewascatalog.org/phecodes}
}
\usage{
diseaseDxIcdMap
}
\description{
This table provides mapping between diseases and ICD9 and ICD10 codes that
represent being diagnosed with them.
}
\seealso{
[]
}
\keyword{datasets}
