% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{diseaseHpoMap}
\alias{diseaseHpoMap}
\title{Mapping of Mendelian diseases and their clinical features}
\format{
A data.table with the following columns:
\itemize{
\item \code{disease_id}: Numeric vector of OMIM disease identifiers
\item \code{disease_name}: Character vector of disease names
\item \code{hpo_term_id}: Character vector of HPO identifiers corresponding to each
disease's clinical features
\item \code{hpo_term_name}: Character vector of HPO term descriptions
}
}
\source{
\url{https://hpo.jax.org/app/download/annotation}
}
\usage{
diseaseHpoMap
}
\description{
This table provides a mapping between Mendelian diseases and their clinical
features, represented as Human Phenotype Ontology (HPO) terms. The mapping is
based on annotations from Online Mendelian Inheritance in Man (OMIM).
}
\seealso{
\code{\link[=mapDiseaseToPhecode]{mapDiseaseToPhecode()}}
}
\keyword{datasets}
