% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{diseaseHpoMap}
\alias{diseaseHpoMap}
\title{Mapping of disease entities and their clinical features}
\format{
A data table with 166,143 rows and the 6 following columns:

\describe{
\item{\code{db_name}}{Character vector indicating the reference database
used for each annotation ('OMIM', 'DECIPHER', 'ORPHA')}
\item{\code{disease_id}}{Numeric vector of the database identifier
corresponding to the disease being annotated}
\item{\code{disease_name}}{Character vector of disease name}
\item{\code{hpo_id}}{Character vector of the HPO identifier for the
phenotype linked to \code{disease_id}}
\item{\code{term_id}}{Numeric vector containing the numeric portion of
\code{hpo_id}}
\item{\code{hpo_term_name}}{Character vector containing the description
corresponding to \code{hpo_id}}
}
}
\source{
\url{https://hpo.jax.org/app/download/annotation}
}
\usage{
diseaseHpoMap
}
\description{
This table provides mapping between disease entities and the clinical
features they are linked to using annotations from multiple
databases (OMIM, Orphanet, and Decipher).
}
\seealso{
\code{\link[=mapDiseaseToPhecode]{mapDiseaseToPhecode()}}
}
\keyword{datasets}
