% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genoSample}
\alias{genoSample}
\title{Sample genotype dataset}
\format{
A data table with 5 rows and the 805 columns. The first column is
\code{person_id} and the rest are variant IDs.
}
\usage{
genoSample
}
\description{
This table contains a sample dataset of genotypes.
}
\keyword{datasets}
