% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_association.R
\name{getGeneticAssociations}
\alias{getGeneticAssociations}
\title{Perform genotype association tests}
\usage{
getGeneticAssociations(
  scores,
  genotypes,
  demos,
  diseaseVariantMap,
  glmFormula,
  modelType = c("additive", "dominant", "recessive", "genotypic"),
  level = 0.95,
  dopar = FALSE
)
}
\arguments{
\item{scores}{A data.table of phenotype risk scores. Must have columns
\code{person_id}, \code{disease_id}, \code{score}.}

\item{genotypes}{A 'BEDMatrix' object linked to the PLINK bed file containing
genetic data. The row names correspond to the \code{person_id}'s in
\code{demos} and \code{scores} tables. The column names correspond to variant IDs.}

\item{demos}{A data.table of covariates to be used in the association
analysis. Must have column \code{person_id}.}

\item{diseaseVariantMap}{A data.table that maps diseases to variants.
Must have columns \code{disease_id} and \code{vid}.
\code{disease_id} must include all disease IDs in \code{scores}.}

\item{glmFormula}{A formula object representing the linear model of covariates
to be used in the genetic association tests. All covariates must be
existing columns in \code{demos}.}

\item{modelType}{Coding of the genotypes. Either 'additive' or 'genotypic'.}

\item{level}{Confidence level used for calculating confidence intervals.
Default is 0.95.}

\item{dopar}{Logical indicating whether to run calculations in parallel if
a parallel backend is already set up, e.g., using
\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}. Recommended to minimize runtime.}
}
\value{
A data.table containing to results of the genotype association tests
with the following columns. If the model fails to converge,
NAs will be reported.
\itemize{
\item \code{disease_id}: disease identifier
\item \code{gene}: name of the gene
\item \code{vid}: variant identifier
\item \code{beta}: The beta coefficient for the association
\item \code{se}: The standard error for the beta coefficient
\item \code{pval}: The p-value for the association
\item \code{ci_lower}: The lower bound of the \code{confint} confidence interval
\item \code{ci_upper}: The upper bound of the \code{confint} confidence interval
}
}
\description{
Perform association tests between the phenotype risk scores and genotypes
}
\examples{
library('data.table')
library('BEDMatrix')

# map ICD codes to phecodes
phecodeOccurrences = getPhecodeOccurrences(icdSample)

# calculate weights
weights = getWeights(demoSample, phecodeOccurrences)

# vector of OMIM disease IDs to calculate PheRS for
diseaseId = 154700

# mape diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()

# calculate PheRS
scores = getScores(
  demoSample, phecodeOccurrences, weights, diseasePhecodeMap[disease_id == diseaseId])

# create a map of diseases and variants
diseaseVariantMap = data.table(disease_id = diseaseId, vid = paste0('snp', 1:10))

# load genetic data
npop = 50
nvar = 10
genoSample = BEDMatrix(system.file('extdata', 'geno_sample.bed', package = 'phers'))
colnames(genoSample) = paste0('snp', 1:nvar)
rownames(genoSample) = 1:npop

# run genetic association tests
genoStats = getGeneticAssociations(
  scores, genoSample, demoSample, diseaseVariantMap, glmFormula = ~ sex,
  modelType = 'additive')

  
}
