% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{getPhecodeOccurrences}
\alias{getPhecodeOccurrences}
\title{Map ICD code occurrences to phecode occurrences}
\usage{
getPhecodeOccurrences(
  icdOccurrences,
  icdPhecodeMap = phers::icdPhecodeMap,
  DxIcd = NULL
)
}
\arguments{
\item{icdOccurrences}{A data.table of occurrences of ICD codes for each
person in the cohort. Must have columns \code{person_id}, \code{icd}, and \code{flag}.}

\item{icdPhecodeMap}{A data.table containing the mapping between ICD codes
and phecodes. Must have columns \code{icd}, \code{phecode}, and \code{flag}. By default uses
the mapping included in this package.}

\item{DxIcd}{character vector of diagnostic ICD codes to remove}
}
\value{
A data.table of phecode occurrences for each person.
}
\description{
This function takes a data table of patient ICD codes and maps them to
phecodes.
}
