% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{getResidualScores}
\alias{getResidualScores}
\title{Calculate residual phenotype risk scores}
\usage{
getResidualScores(demos, scores, glmFormula, dopar = FALSE)
}
\arguments{
\item{demos}{A data.table of covariates to be used in the calculation
of residual phenotype risk score. Must have column \code{person_id}.}

\item{scores}{A data.table containing the phenotype risk score for each person for
each disease. Must have columns \code{person_id}, \code{disease_id}, and \code{score}.}

\item{glmFormula}{A formula object representing the linear model of covariates
to be used in the calculation of residual phenotype risk score.
All covariates must be existing columns in \code{demos}.}

\item{dopar}{Logical indicating whether to run calculations in parallel if
a parallel backend is already set up, e.g., using
\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}. Recommended to minimize runtime.}
}
\value{
A data.table containing the phenotype risk score and the residual
phenotype risk score for each person for each disease.
}
\description{
The residual phenotype risk score corresponds to how much a person's
phenotype risk score deviates from what is expected given
their characteristics (e.g. age, sex, and race)
}
