% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{getScores}
\alias{getScores}
\title{Calculate phenotype risk scores}
\usage{
getScores(demos, phecodeOccurrences, weights, diseasePhecodeMap)
}
\arguments{
\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}.}

\item{phecodeOccurrences}{A data.table of occurrences of phecodes for each
person in the cohort. Must have columns \code{person_id} and \code{phecode}.}

\item{weights}{A data.table where each row is a phecode and the weight
corresponding to it. The columns are \code{phecode} and \code{w}.}

\item{diseasePhecodeMap}{A data.table containing the mapping between
diseases and phecodes. Must have columns \code{disease_id} and \code{phecode}.}
}
\value{
A data.table containing the phenotype risk score for each person for
each disease.
}
\description{
A person's phenotype risk score for a particular disease corresponds to the
sum of the weights of the disease-relevant phecodes that the person has
received.
}
\examples{
library('data.table')

# map ICD codes to phecodes
phecodeOccurrences = getPhecodeOccurrences(icdSample)

# calculate weights
weights = getWeights(demoSample, phecodeOccurrences)

# vector of OMIM disease IDs to calculate PheRS for
diseaseIds = c(154700)

# mape diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()

# calculate PheRS
scores = getScores(
  demoSample, phecodeOccurrences, weights, diseasePhecodeMap[disease_id \%in\% diseaseIds])

# calculate residual PheRS
rscores = getResidualScores(demoSample, scores, glmFormula = as.formula(~ sex))

  
}
