% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{getWeights}
\alias{getWeights}
\title{Calculate phecode-specific weights for phenotype risk scores}
\usage{
getWeights(
  demos,
  phecodeOccurrences,
  method = c("prevalence", "logistic", "cox", "loglinear"),
  methodFormula = NULL,
  dopar = FALSE
)
}
\arguments{
\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}.}

\item{phecodeOccurrences}{A data.table of phecode occurrences for each person
in the cohort. Must have columns \code{person_id} and \code{phecode}.}

\item{method}{A string indicating the statistical model for calculating weights.}

\item{methodFormula}{A formula representing the right hand side of the model
corresponding to \code{method}. All terms in the formula must correspond to
columns in \code{demos}.}

\item{dopar}{Logical indicating whether to run calculations in parallel if
a parallel backend is already set up, e.g., using
\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}. Recommended to minimize runtime.}
}
\value{
A data.table with columns \code{phecode}, \code{prev} (prevalence), and \code{w}
(weight) under the \code{prevalence} method and columns \code{person_id}, \code{phecode},
\code{prob} (probability), and \code{w} under \code{logistic}, \code{cox}, and \code{loglinear}
methods. Prevalence corresponds to fraction of the cohort that has at
least one occurrence of the given phecode. Probability corresponds to
parametric probability of having a phecode for each individual calculated
based on the selected \code{method} and \code{methodFormula}. Weight is calculated
as \code{-log10} prevalence or probability.
}
\description{
This is typically the second step of an analysis using phenotype risk scores,
the next is \code{\link[=getScores]{getScores()}}.
}
\examples{
library('data.table')

# map ICD codes to phecodes
phecodeOccurrences = getPhecodeOccurrences(icdSample)

# calculate weights
weights = getWeights(demoSample, phecodeOccurrences)

# OMIM disease IDs for which to calculate phenotype risk scores
diseaseId = 154700

# map diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()

# calculate scores
scores = getScores(
  demoSample, phecodeOccurrences, weights, diseasePhecodeMap[disease_id == diseaseId])

# calculate residual scores
rscores = getResidualScores(demoSample, scores, lmFormula = ~ sex)
}
\seealso{
\code{\link[=getPhecodeOccurrences]{getPhecodeOccurrences()}}, \code{\link[=getScores]{getScores()}}, \code{\link[=phers]{phers()}}
}
