% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phers.R
\name{getWeights}
\alias{getWeights}
\title{Calculate phecode-specific weights for phenotype risk scores}
\usage{
getWeights(demos, phecodeOccurrences)
}
\arguments{
\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}.}

\item{phecodeOccurrences}{A data.table of occurrences of phecodes for each
person in the cohort. Must have columns \code{person_id} and \code{phecode}.}
}
\value{
A data.table containing the prevalence (\code{prev}) and weight (\code{w}) for
each phecode.
}
\description{
The weights correspond to the -log10 prevalences in the cohort.
}
