% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{getWeights}
\alias{getWeights}
\title{Calculate phecode-specific weights for phenotype risk scores}
\usage{
getWeights(
  demos,
  phecodeOccurrences,
  method = c("prevalence", "logistic", "cox", "loglinear"),
  methodFormula = NULL,
  dopar = FALSE
)
}
\arguments{
\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}. When the \code{cox} method is used, \code{demos}
must have columns \code{first_age} and \code{last_age} corresponding to first and
last age of visit (in years).}

\item{phecodeOccurrences}{A data.table of phecode occurrences for each person
in the cohort. Must have columns \code{person_id} and \code{phecode} under the
"prevalence" or "logistic" methods, columns \code{person_id}, \code{phecode}, and
\code{num_occurrences} under the "loglinear" method, and columns \code{person_id},
\code{phecode}, and \code{occurrence_age} under the "cox" method. \code{num_occurrences}
refers to the number of unique dates a phecode was recorded for a person.
\code{occurrence_age} refers to the first age (in years) a person acquired a
phecode.}

\item{method}{A string indicating the statistical model for calculating
weights.}

\item{methodFormula}{A formula representing the right-hand side of the model
corresponding to \code{method}. All terms in the formula must correspond to
columns in \code{demos}. Do not use age-related covariates with the "cox"
method.}

\item{dopar}{Logical indicating whether to run calculations in parallel if
a parallel backend is already set up, e.g., using
\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}. Recommended to minimize runtime.}
}
\value{
A data.table with various columns. If \code{method} is "prevalence":
\code{phecode}, \code{prev} (prevalence), and \code{w} (weight). If \code{method} is
"logistic" or "cox": \code{person_id}, \code{phecode}, \code{prob} (probability), and \code{w}.
If \code{method} is "loglinear": \code{person_id}, \code{phecode}, and \code{w}. Prevalence
corresponds to fraction of the cohort that has at least one occurrence of
the given phecode. Probability corresponds to predicted probability of
given individual having a given phecode based on \code{method} and
\code{methodFormula}. For the "prevalence" \code{method}, weight is calculated as
\code{-log10(prev)}, for "logistic" and "cox" as \code{-log10(prob)}, and for
"loglinear" as the difference between observed and predicted
\code{log2(num_occurrences + 1)}.
}
\description{
This is typically the second step of an analysis using phenotype risk scores,
the next is \code{\link[=getScores]{getScores()}}.
}
\examples{
library('data.table')
library('survival')

# map ICD codes to phecodes
phecodeOccurrences = getPhecodeOccurrences(icdSample)

# calculate weights using the prevalence method
weightsPrev = getWeights(demoSample, phecodeOccurrences)

# calculate weights using the logistic method
weightsLogistic = getWeights(
  demoSample, phecodeOccurrences, method = 'logistic', methodFormula = ~ sex)

# calculate weights using the loglinear method
phecodeOccurrences2 = phecodeOccurrences[, .(
  num_occurrences = uniqueN(entry_date)), by = .(person_id, phecode)]
weightsLoglinear = getWeights(
  demoSample, phecodeOccurrences2, method = 'loglinear', methodFormula = ~ sex)

# calculate weights using the cox method
phecodeOccurrences3 = phecodeOccurrences[, .(
  first_occurrence_date = min(entry_date)) , by = .(person_id, phecode)]
phecodeOccurrences3 = merge(
  phecodeOccurrences3, demoSample[, .(person_id, dob)], by = 'person_id')
phecodeOccurrences3[,
  occurrence_age := as.numeric((first_occurrence_date - dob)/365.25)]
phecodeOccurrences3[, `:=`(first_occurrence_date = NULL, dob = NULL)]
demoSample3 = demoSample[, .(
  person_id, sex,
  first_age = as.numeric((first_visit_date - dob)/365.25),
  last_age = as.numeric((last_visit_date - dob)/365.25))]

weightsCox = getWeights(
  demoSample3, phecodeOccurrences3, method = 'cox', methodFormula = ~ sex)
}
\seealso{
\code{\link[=getPhecodeOccurrences]{getPhecodeOccurrences()}}, \code{\link[=getScores]{getScores()}}, \code{\link[=phers]{phers()}}
}
