% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{hpoPhecodeMap}
\alias{hpoPhecodeMap}
\title{Mapping of HPO terms and phecodes}
\format{
A data.table with the following columns:
\itemize{
\item \code{hpo_term_id}: Character vector of HPO term identifiers
\item \code{hpo_term_name}: Character vector of HPO term descriptions
\item \code{phecode}: Character vector of phecodes
\item \code{phecode_name}: Character vector of phecode descriptions
}
}
\usage{
hpoPhecodeMap
}
\description{
This table provides a mapping between Human Phenotype Ontology (HPO)
terms and phecodes, and is useful for using phecodes to represent the
clinical features of Mendelian diseases (version 1.2).
}
\seealso{
\code{\link[=mapDiseaseToPhecode]{mapDiseaseToPhecode()}}
}
\keyword{datasets}
