% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HPOPhecodeMap}
\alias{HPOPhecodeMap}
\title{Mapping of HPO terms and phecodes}
\format{
A data table with 14,160 rows and the 4 following columns:

\describe{
\item{\code{term_id}}{Numeric vector representing the HPO term ID to be
mapped}
\item{\code{hpo_term_name}}{Character vector containing the description
corresponding to each HPO term ID}
\item{\code{phecode}}{Character vector representing the phecode mapping}
\item{\code{phecode_string}}{Character vector containing the description
corresponding to each phecode}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6857501/#sup1}
}
\usage{
HPOPhecodeMap
}
\description{
This table provides mapping between Human Phenotype Ontology (HPO)
annotations and phecodes (groupings of ICD-9-CM and ICD-10-CM)
}
\seealso{
\code{\link[=mapDiseaseToPhecode]{mapDiseaseToPhecode()}}
}
\keyword{datasets}
