% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{icdPhecodeMap}
\alias{icdPhecodeMap}
\title{Mapping of ICD codes and phecodes}
\format{
A data.table with the following columns:
\itemize{
\item \code{icd}: Character vector of ICD codes
\item \code{flag}: Integer vector of the vocabulary of the ICD code
(\strong{9}: ICD-9-CM,  \strong{10}: ICD-10-CM)
\item \code{icd_name}: Character vector of ICD code descriptions
\item \code{phecode}: Character vector of phecodes
\item \code{phecode_name}: Character vector of phecode descriptions
}
}
\source{
\url{https://phewascatalog.org/phecodes}
}
\usage{
icdPhecodeMap
}
\description{
This table provides a mapping between International Classification of
Diseases 9th and 10th revisions (ICD-9-CM and ICD-10-CM) and phecodes
(version 1.2).
}
\seealso{
\code{\link[=getPhecodeOccurrences]{getPhecodeOccurrences()}}
}
\keyword{datasets}
