% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{icdPhecodeMap}
\alias{icdPhecodeMap}
\title{Mapping of ICD and phecodes}
\format{
A data table with 156,641 rows and the 3 following columns:

\describe{
\item{\code{icd}}{Character vector representing the ICD code to be mapped}
\item{\code{phecode}}{Character vector representing the phecode mapping}
\item{\code{flag}}{Integer vector representing the vocabulary of the
ICD code (\strong{9}: ICD-9-CM,  \strong{10}: ICD-10-CM)}
}
}
\source{
\url{https://phewascatalog.org/phecodes_v1_1}
}
\usage{
icdPhecodeMap
}
\description{
This table provides mapping between International Classification of
Diseases 9th and 10th revisions (ICD-9-CM and ICD-10-CM) and phecodes
(groupings of ICD codes).
}
\seealso{
\code{\link[=getPhecodeOccurrences]{getPhecodeOccurrences()}}
}
\keyword{datasets}
