% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{icdSample}
\alias{icdSample}
\title{Sample table of ICD occurrences}
\format{
A data.table with the following columns:
\itemize{
\item \code{person_id}: Character vector of the identifier for each person
\item \code{icd}: Character vector of the ICD codes recorded for each person
\item \code{flag}: Integer vector of the vocabulary of the ICD code
(\strong{9}: ICD-9-CM, \strong{10}: ICD-10-CM)
\item \code{entry_date}: Vector of type \code{Date} indicating the date each ICD code was
recorded.
}
}
\usage{
icdSample
}
\description{
The data are artificial and do not correspond to real patients.
}
\seealso{
\code{\link[=getPhecodeOccurrences]{getPhecodeOccurrences()}}, \code{\link[=getWeights]{getWeights()}}, \code{\link[=getScores]{getScores()}}
}
\keyword{datasets}
