% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icdSample}
\alias{icdSample}
\title{Sample diagnostic code dataset}
\format{
A data table with 9 rows and the 3 following columns:

\describe{
\item{\code{person_id}}{Character vector of the identifier for each
person}
\item{\code{icd}}{Character vector containing the ICD codes recorded
for each person}
\item{\code{flag}}{Numeric vector representing the vocabulary of the
ICD code (\strong{9}: ICD-9-CM,  \strong{10}: ICD-10-CM)}
}
}
\usage{
icdSample
}
\description{
This table contains a sample dataset of patient diagnostic codes
(ICD-9-CM and ICD-10-CM).
}
\seealso{
\code{\link[=calcWeights]{calcWeights()}}, \code{\link[=calcPheRS]{calcPheRS()}}
}
\keyword{datasets}
