% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_disease_to_phecode.R
\name{mapDiseaseToPhecode}
\alias{mapDiseaseToPhecode}
\title{Map diseases IDs to phecodes using HPO terms}
\usage{
mapDiseaseToPhecode(
  diseaseIDs,
  dbName = "OMIM",
  diseaseHPOMap = phers::diseaseHPOMap,
  HPOPhecodeMap = phers::HPOPhecodeMap
)
}
\value{
A data.table containing the mapping between diseases and
their phenotypes. The columns are \code{disease_ID}, \code{phecode}.
}
\description{
Map diseases IDs to phecodes using HPO terms
}
