% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapDiseaseToPhecode}
\alias{mapDiseaseToPhecode}
\title{Map diseases entities to phecodes using HPO terms}
\usage{
mapDiseaseToPhecode(
  diseaseIds = unique(phers::diseaseHpoMap$disease_id),
  diseaseHpoMap = phers::diseaseHpoMap,
  hpoPhecodeMap = phers::hpoPhecodeMap
)
}
\arguments{
\item{diseaseIds}{A numeric vector of OMIM disease identifiers to be mapped.}

\item{diseaseHpoMap}{A data.table containing the mapping between disease
entities in \code{diseaseIds} and HPO terms. Must have columns \code{disease_id} and
\code{term_id}. By default uses the map included in this package.}

\item{hpoPhecodeMap}{A data.table containing the mapping between HPO terms
and phecodes. Must have columns \code{term_id} and \code{phecode}. By default uses the
map included in this package.}
}
\value{
A data.table containing the mapping between diseases and phecodes.
The columns are \code{disease_id} and \code{phecode}.
}
\description{
This function takes a list of disease identifiers and returns the
clinical features mapped to them as phecodes.
}
