% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{preCalcWeights}
\alias{preCalcWeights}
\title{pre-calculated weights}
\format{
A data table with 1865 rows and the 3 following columns:
\itemize{
\item \code{phecode}: Character vector of phecodes
\item \code{prev}: Prevelance of each phecode within SD
\item \code{w}: Weight calculated for each phecode using the PheRS model
}
}
\source{
<>
}
\usage{
preCalcWeights
}
\description{
This table provides weights for 1865 phecodes calculated using phecode Version
1.2 and data from Vanderbilt University Medical Center Synthetic Derivative (SD)
}
\seealso{
\code{\link[=getWeights]{getWeights()}}
}
\keyword{datasets}
