% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_association.R
\name{runLinear}
\alias{runLinear}
\title{Perform a single genotype association test}
\usage{
runLinear(lmInput, glmFormula, modelType = "additive")
}
\arguments{
\item{lmInput}{A data.table of of the input to the genetic association model.
Must have columns \code{score}, \code{allele_count} and any other covariates to be used
in the linear regression model. \code{allele_count} should contain the number of
copies of the minor allele each person has (additive model).}

\item{glmFormula}{a formula object representing the linear model of covariates
to be used in the genetic association test. All covariates must be existing
columns in \code{lmInput}.}

\item{modelType}{Coding of the genotypes. Either 'additive' or 'genotypic'}
}
\value{
A data.table containing to results of the genotype association test
with the following columns. If the model fails to converge,
NAs will be reported.
\itemize{
\item \code{beta}: The beta coefficient for the variant
\item \code{se}: The standard error for the beta coefficient
\item \code{p}: The p-value for the variant
\item \code{lower}: The lower bound of the \code{confint} confidence interval
\item \code{upper}: The upper bound of the \code{confint} confidence interval
}
}
\description{
Perform an association test between one variant and the phenotype risk scores
corresponding to one disease.
}
